Upgrade Buttplug to latest version & add WebBluetooth check
Fixed the broken buttplug.io intergration by upgrading to the latest
stable version and changing any of the API calls that were affected by
the upgrade.

To make the latest version of buttplug.io work I had to introduce
`react-app-rewire` so we can enforce additional webpack config to load
the _web_ version of buttplug rather than the node version.

Additionally introduced a guard on the vibrator settings dialog to first
check if WebBluetooth is actually available first before offering to
connect a device to prevent misleading end users.

Fixes: #2